+++ next-docs/pages/templates/sportsbet/vip.tsx
@@ -1,9 +1,8 @@
-import React, { ReactNode } from 'react';
-import { Carousel } from '@heathmont/moon-components';
-import { Button } from '@heathmont/moon-core-tw';
-import { MailEnvelope, ArrowsLeft, ShopCrypto } from '@heathmont/moon-icons';
-
+import React, { ReactNode, useEffect, useState } from 'react';
+import { Button, Accordion } from '@heathmont/moon-core-tw';
+import ControlsChevronDown from '@heathmont/moon-icons-tw/lib/icons/ControlsChevronDown';
 import Image from 'next/image';
+import image1 from '../../../public/templates/bitcasino/image1.png';
 import explore1 from '../../../public/templates/explore-1.png';
 import explore2 from '../../../public/templates/explore-2.png';
 import explore3 from '../../../public/templates/explore-3.png';
@@ -11,8 +10,49 @@ import explore4 from '../../../public/templates/explore-4.png';
 import gift from '../../../public/templates/gift.png';
 import handshake from '../../../public/templates/handshake.png';
 import promo from '../../../public/templates/promo.png';
+import Plus18 from '../../../public/templates/sportsbet/18plus.png';
+import Award1 from '../../../public/templates/sportsbet/award1.png';
+import Award2 from '../../../public/templates/sportsbet/award2.png';
+import Award3 from '../../../public/templates/sportsbet/award3.png';
+import Curacao from '../../../public/templates/sportsbet/curacao.png';
 import SportsbetLogo from '../../../public/templates/sportsbet/logo.png';
+import SbcAwards from '../../../public/templates/sportsbet/sbc-awards.png';
 import support from '../../../public/templates/support.png';
+import CTA from "../components/cta";
+import Experience from "../components/experience";
+import Explore from "../components/explore";
+import Family from "../components/family";
+import Header from "../components/header";
+import Headline from "../components/headline";
+import Hero from "../components/hero";
+import Social from "../components/social";
+import Subtitle from "../components/subtitle";
+import Arsenal from '../icons/arsenal';
+import Bitcasino from '../icons/bitcasino';
+import BrazilianIc from '../icons/brazilianIc';
+import Cardano from '../icons/cardano';
+import EthereumIc from '../icons/ethereumIc';
+import EuroIcon from '../icons/euro';
+import Flamengo from '../icons/flamengo';
+import IndianIc from '../icons/indianIc';
+import Instagram from '../icons/instagram';
+import JapanesIc from '../icons/japanesIc';
+import LtchLogo from '../icons/ltchLogo';
+import Mastercard from '../icons/mastercard';
+import Neteller from '../icons/neteller';
+import PlusFive from '../icons/plusFive';
+import Scril from '../icons/skrill';
+import Southampton from '../icons/southampton';
+import Telegram from '../icons/telegram';
+import Tiktok from '../icons/tiktok';
+import TronLogo from '../icons/tronLogo';
+import Twitter from '../icons/twitter';
+import UsdIc from '../icons/usdtIc';
+import Visa from '../icons/visa';
+import Watford from '../icons/watford';
+import XAbove from '../icons/xAbove';
+import Youtube from '../icons/youtube';
+
 
 interface ExperienceCarousel {
   index?: number;
@@ -29,7 +69,7 @@ const ExperienceCarouselItem = ({
 }: ExperienceCarousel) => {
   return (
     <div
-      className={`w-[272px] h-full bg-[#2C323A] rounded-lg flex flex-col p-6 sm:my-8 sm:h-auto ${
+      className={`w-[285px] h-full bg-goku rounded-lg flex flex-col p-6 sm:my-8 sm:h-auto ${
         index && index % 2 ? 'sm:mt-[40%]' : ''
       }`}
     >
@@ -54,20 +94,48 @@ const ExploreCarouselItem = ({
 }: {
   items: Array<{ image: any; title: string }>;
 }) => {
+  const [itemWidth, setItemWidth] = useState(160);
+  const [itemHeight, setItemHeight] = useState(200);
+
+  useEffect(() => {
+    const resizeItems = () => {
+      if (window?.innerWidth) {
+        setItemWidth(window.innerWidth <= 1024 ? 160 : 280);
+        setItemHeight(window.innerWidth <= 1024 ? 180 : 317);
+      }
+    };
+
+    resizeItems();
+
+    window.addEventListener("resize", resizeItems);
+
+    return () => {
+      window.removeEventListener("resize", resizeItems);
+    };
+  }, []);
+
   return (
     <div className="w-full h-full flex p-6">
-      <div className="w-[168px] me-3 flex flex-col items-stretch">
+      <div className="w-[160px] me-4 flex flex-col items-stretch lg:w-[280px]">
         <div className="h-16" />
-        <div className="grow-1 mb-2 relative rounded-lg">
-          <Image src={items[0].image} />
+        <div className="w-[160px] h-[180px] grow-1 mb-4 relative rounded-lg lg:w-[280px] lg:h-[317px]">
+          <Image
+            src={items[0].image}
+            width={itemWidth}
+            height={itemHeight}
+          />
           <div className="absolute top-0 left-0 opacity-0 w-full h-full flex items-end pointer hover:opacity-100 bg-[#0000008F]">
             <p className="text-moon-18 font-semibold m-3 w-full">
               {items[0].title}
             </p>
           </div>
         </div>
-        <div className="grow-1 relative rounded-lg">
-          <Image src={items[1].image} />
+        <div className="w-[160px] h-[180px] grow-1 relative rounded-lg lg:w-[280px] lg:h-[317px]">
+          <Image
+            src={items[1].image}
+            width={itemWidth}
+            height={itemHeight}
+          />
           <div className="absolute top-0 left-0 opacity-0 w-full h-full flex items-end pointer hover:opacity-100 bg-[#0000008F]">
             <p className="text-moon-18 font-semibold m-3 w-full">
               {items[1].title}
@@ -75,17 +143,25 @@ const ExploreCarouselItem = ({
           </div>
         </div>
       </div>
-      <div className="w-[168px] flex flex-col items-stretch relative">
-        <div className="grow-1 mb-2 relative rounded-lg">
-          <Image src={items[2].image} />
+      <div className="w-[160px] me-5 flex flex-col items-stretch lg:w-[280px]">
+        <div className="w-[160px] h-[180px] grow-1 mb-4 relative rounded-lg lg:w-[280px] lg:h-[317px]">
+          <Image
+            src={items[2].image}
+            width={itemWidth}
+            height={itemHeight}
+          />
           <div className="absolute top-0 left-0 opacity-0 w-full h-full flex items-end pointer hover:opacity-100 bg-[#0000008F]">
             <p className="text-moon-18 font-semibold m-3 w-full">
               {items[2].title}
             </p>
           </div>
         </div>
-        <div className="grow-1 relative rounded-lg">
-          <Image src={items[3].image} />
+        <div className="w-[160px] h-[180px] grow-1 relative rounded-lg lg:w-[280px] lg:h-[317px]">
+          <Image
+            src={items[3].image}
+            width={itemWidth}
+            height={itemHeight}
+          />
           <div className="absolute top-0 left-0 opacity-0 w-full h-full flex items-end pointer hover:opacity-100 bg-[#0000008F]">
             <p className="text-moon-18 font-semibold m-3 w-full">
               {items[3].title}
@@ -98,11 +174,7 @@ const ExploreCarouselItem = ({
   );
 };
 
-const Headline = ({ text }: { text: string }) => (
-  <p className="text-[32px] text-center font-semibold p-4 lg:text-[64px] lg:leading-[3.5rem] sm:text-left sm:text-moon-48">
-    {text}
-  </p>
-);
+
 const Title = ({ text, classes }: { text: string; classes?: string }) => (
   <p
     className={`text-[32px] text-center font-semibold p-4 lg:text-moon-48 sm:text-moon-32 ${
@@ -112,17 +184,9 @@ const Title = ({ text, classes }: { text: string; classes?: string }) => (
     {text}
   </p>
 );
-const Subtitle = ({ text, classes }: { text: string; classes?: string }) => (
-  <p
-    className={`text-[16px] text-center px-3 pt-2 lg:text-moon-24 sm:text-moon-16 ${
-      classes ?? ''
-    }`}
-  >
-    {text}
-  </p>
-);
 
 const Vip = () => {
+  const [isScrolled, setIsScrolled] = useState(false);
   const experienceItems = [
     {
       icon: handshake,
@@ -166,22 +230,6 @@ const Vip = () => {
       title: '24/7 priority support',
       points: ['No queues, no waiting.', 'We attend to your needs first.'],
     },
-    {
-      icon: gift,
-      title: 'Luxurious holiday gifts',
-      points: [
-        'High-value gifts on your birthday and your anniversary.',
-        'Personalised seasonal offers.',
-      ],
-    },
-    {
-      icon: promo,
-      title: 'VIP exclusive promos',
-      points: [
-        'The best betting experiences are reserved for you.',
-        'Only visible to VIP members.',
-      ],
-    },
   ];
   const exploreItems = [
     [
@@ -222,197 +270,492 @@ const Vip = () => {
     ],
   ];
 
+  useEffect(() => {
+    const toggleHeaderBackground = (e: Event) => {
+      setIsScrolled(window.scrollY >= 30);
+    };
+
+    window.addEventListener('scroll', toggleHeaderBackground);
+
+    return () => {
+      window.removeEventListener('scroll', toggleHeaderBackground);
+    };
+  }, []);
+
   return (
     <div className="flex flex-col grow overflow-hidden">
-      <div className="fixed top-0 left-0 w-full flex items-center justify-center my-5 mx-4 z-10">
-        <div className="absolute left-0">
-          <ArrowsLeft fontSize="2rem" />
-        </div>
-        <Image src={SportsbetLogo} />
-      </div>
-      <div className="mt-20 sm:flex sm:flex-row-reverse sm:items-stretch">
-        <div className="flex flex-col h-[80vw] sm:items-center sm:w-[50%] sm:h-auto sm:min-h-[30vw] md:min-h-[25vw]">
-          <div className="relative w-full grow">
-            <div
-              className="w-full h-full bg-contain bg-left bg-no-repeat"
-              style={{
-                backgroundImage: 'url(/templates/sportsbet/ambasadors.png)',
-              }}
-            />
-            <div
-              className="absolute left-0 bottom-0 w-full h-[150px]"
-              style={{
-                background:
-                  'linear-gradient(180deg, rgba(26, 33, 42, 0) 0%, rgba(26, 33, 42, 0.479167) 47.92%, rgba(26, 33, 42, 0.81) 67.19%, #1A212A 82.29%, #1A212A 100%)',
-              }}
-            />
-          </div>
-        </div>
-        <div className="flex flex-col items-end relative -top-[30vw] px-8 sm:py-20 sm:w-[50%] sm:top-0 sm:items-center md:ps-[15%]">
-          <div className="flex flex-col max-w-full sm:w-[550px]">
-            <Headline text="Bitcasino VIP: For money can’t buy experiences!" />
-            <Subtitle
-              text="Join the VIP Bitcasino Club - An exclusive world of crypto leaders
+      <Header
+        logo={(<div className="mt-2"><Image src={SportsbetLogo} /></div>)}
+        height={40}
+        isScrolled={isScrolled}
+      />
+      <Hero
+        headline={(<Headline
+          text="Bitcasino VIP: For money can’t buy experiences!"
+          classes="sm:px-0"
+        />)}
+        subtitle={(<Subtitle
+          classes="items-center mt-3 sm:px-0 sm:text-left sm:text-[18px] sm:leading-6 md:text-[24px] md:leading-8"
+          text=" Join the VIP Bitcasino Club - An exclusive world of crypto leaders
               and high-rollers who dare to take risks and enjoy life to its
               fullest."
-              classes="sm:text-left"
-            />
-            <Button
-              variant="primary"
-              className="btn-primary py-3 px-4 gap-1 text-moon-16 rounded-moon-i-sm mx-3 my-6 sm:self-start"
-            >
-              Sign up now
-            </Button>
-          </div>
-        </div>
-      </div>
-      <div className="flex flex-col items-center self-center sm:w-[70%]">
-        <div className="flex flex-col relative -top-[100px] my-8 sm:top-0 sm:mb-0">
-          <Title text="A VIP experience tailored to your liking" />
-          <Subtitle
+        />)}
+        button={(<Button
+          variant="primary"
+          width={124}
+          height={48}
+          className="btn-primary py-3 px-4 gap-1 text-moon-16 rounded-moon-i-sm mx-3 my-6 sm:mx-0 sm:px-0 sm:self-start w-full sm:w-[124px] "
+        >
+          Sign up now
+        </Button>)}
+        heroImage={(<div
+          className="w-full h-full bg-contain bg-bottom bg-no-repeat min-w-[100vw] min-h-[40vh] sm:min-w-[44vw] sm:min-h-[auto]"
+          style={{
+            backgroundImage: 'url(/templates/sportsbet/ambasadors.png)',
+          }}
+        />)}
+        bottomGradient='linear-gradient(180deg, rgba(26, 33, 42, 0) 0%, rgba(26, 33, 42, 0.479167) 47.92%, rgba(26, 33, 42, 0.81) 67.19%, #1A212A 82.29%, #1A212A 100%)'
+      />
+
+      <div className="flex flex-col items-center self-center mt-[200px] sm:w-[80%] sm:mt-[124px]">
+        <Experience
+          title={(<Title
+            text="A VIP experience tailored to your liking"
+            classes="w-[380px] leading-10 mb-4 sm:mb-0 md:w-[944px] sm:w-[600px] sm:leading-0"
+          />)}
+          subtitle={(<Subtitle
+            classes="w-[348px] p-5 md:w-[700px] sm:w-[650px] sm:p-0 sm:pt-4"
             text="Enjoy exclusive bonuses and invites to prestigious events for free.
-            Get all your indulgences taken care of."
-          />
-          <div className="w-full flex max-w-[100vw] items-center justify-center px-5 py-8 xl:min-w-[1190px] lg:min-w-[900px] sm:mt-12 sm:max-w-[70vw]">
-            <Carousel
-              items={experienceItems.map((item, index) => (
-                <ExperienceCarouselItem
-                  title={item.title}
-                  points={item.points}
-                  icon={item.icon}
-                  index={index}
-                />
-              ))}
-            />
-          </div>
-        </div>
-        <div className="flex flex-col relative -top-[90px] sm:flex-row-reverse sm:top-0 sm:flex-wrap sm:justify-center">
-          <div className="flex flex-col sm:w-[320px] sm:justify-center sm:mb-20">
-            <Title
+              Get all your indulgences taken care of."
+          />)}
+          carouselItems={experienceItems}
+          carouselItemBackground='goku'
+        />
+
+        <div className='mt-[6] sm:mt-[70px] lg:mt-[200px]'>
+          <Explore
+            title={(<Title
               text="Explore the world in true VIP style!"
-              classes="sm:text-left"
-            />
-            <Subtitle
-              classes="sm:text-left"
+              classes="mb-4 text-center leading-10 sm:pl-0 sm:leading-0 xl:text-left"
+            />)}
+            subtitle={(<Subtitle
+              classes="text-center px-5 sm:px-0 xl:text-left"
               text="Enjoy life at its fullest and create great memories as a Bitcasino VIP. We’ll make your dreams come true with money-can't-buy  experiences you’ll never forget. At Bitcasino, we know how indulge you!"
-            />
-          </div>
-          <div className="w-full max-w-[100vw] flex items-center justify-center px-5 py-8 sm:w-[430px]">
-            <Carousel
-              items={exploreItems.map((items) => (
-                <ExploreCarouselItem items={items} />
-              ))}
-            />
-          </div>
-        </div>
-        <div className="flex flex-col relative -top-[50px] sm:flex-row-reverse sm:mt-14 sm:top-0 sm:flex-wrap sm:justify-center">
-          <div className="flex flex-col self-stretch align-center pb-8 sm:ps-12 sm:w-[350px] sm:justify-center lg:w-[550px]">
-            <Title text="VIP runs in the family" classes="sm:text-left" />
-            <Subtitle
-              classes="sm:text-left"
-              text="Sportsbet has legendary athletes in our VIP club. Join our high
-              profile community of crypto leaders and high-rollers!"
-            />
-          </div>
-          <div
-            className="w-full h-full max-w-[100vw] bg-contain bg-center bg-no-repeat h-[80vw] sm:w-[250px] sm:h-[200px] lg:w-[400px] lg:h-[300px]"
-            style={{
-              backgroundImage: 'url(/templates/sportsbet/partnership.png)',
-            }}
+            />)}
+            carouselItems={exploreItems}
+            carouselItemBackground='goku'
           />
         </div>
-        <div className="flex flex-col justify-center items-center sm:mt-16 w-[80vw] sm:w-[40vw]">
-          <Title text="Ready to step up the game?" />
-          <Subtitle
+        <Family
+          title={(<Title
+            text="VIP runs in the family"
+            classes="mb-4 leading-10 sm:text-center md:text-left md:self-start lg:w-[510px]"
+          />)}
+          subtitle={(<Subtitle
+            classes="text-center w-[348px] px-5 pb-[32px] sm:pb-0 sm:w-[100%] md:text-left md:self-start lg:w-[510px]"
+            text="Sportsbet has legendary athletes in our VIP club. Join our high
+              profile community of crypto leaders and high-rollers!"
+          />)}
+          image='/templates/sportsbet/partnership.png'
+        />
+        <CTA
+          title={(<Title
+            text="Ready to step up the game?"
+            classes="mb-4 p-5 leading-10"
+          />)}
+          subtitle={(<Subtitle
             text="Entry into our VIP Club is by invitation only. Get in touch with us
             to see what a tailored VIP package could look like for you. Join the
             family now!"
-            classes="w-[80%]"
+            classes="w-[100%] mb-4 p-5"
+          />)}
+          handleSignup={() => console.log('SIGN UP clicked')}
+          handleContact={() => console.log('CONTACT clicked')}
+        />
+        <div className='mt-[70px] sm:mt-[88px]'>
+          <Social
+            title='Join us and talk to us'
+            onClick={(social: string) => console.log(`Social link - ${social}`)}
           />
-          <div className="flex flex-col my-6 sm:flex-row sm:justify-center sm:items-center">
+        </div>
+      </div>
+      <div className='flex flex-col m-auto sm:w-[80%] sm:mt-[124px] lg:w-[90%] xl:w-[1232px]'>
+        <div
+          className="flex flex-col items-center self-center w-full h-full px-5 mt-16"
+          style={{ background: 'linear-gradient rgba(36, 3, 63, 0) 0% #24033F 100%' }}
+        >
+          <div className="flex flex-col lg:items-start">
+            <p className="text-[24px] text-bulma !text-left leading-8 font-semibold sm:px-3 sm:text-left">
+              Enjoy BTC gaming at the best Bitcoin casino
+            </p>
+            <p className="text-[18px] text-bulma opacity-60 !text-left !leading-8 my-2 sm:px-3 sm:text-left md:text-moon-16 lg:text-moon-16">
+              The future of gaming is here, we are the number one bitcoin casino in 2021 and into 2022.
+              Our range of casino-slots consistently make the top ten across the bitcoin casino world. Read more.
+            </p>
+          </div>
+          <div className="mt-3 mb-3">
             <Button
-              variant="primary"
-              className="btn-primary py-3 px-9 gap-1 text-moon-16 rounded-moon-i-sm mx-1 my-2 sm:self-start"
-            >
-              Sign up
-            </Button>
-
-            <Button
-              variant="primary"
-              className="flex justify-center items-center btn-tertiary py-3 px-5 gap-1 text-moon-16 rounded-moon-i-sm mx-1 my-2 sm:justify-between sm:self-start"
-              icon="left"
-              iconLeft={<MailEnvelope />}
-            >
-              Contact us
-            </Button>
+              className="flex justify-center items-center cursor-pointer bg-goku w-[32px] h-[32px] rounded-full"
+              onClick={() => console.log('Read More')}
+            ><ControlsChevronDown /></Button>
           </div>
         </div>
-        <div className="flex flex-col my-12">
-          <p className="text-moon-18 text-center font-semibold p-4">
-            Join us and talk to us
-          </p>
-          <div className="flex gap-2">
-            <Button
-              iconOnly
-              onClick={() => console.log('Telegram link')}
-              style={{
-                borderRadius: '50%',
-                backgroundColor: '#27A4DB',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
-            <Button
-              iconOnly
-              onClick={() => console.log('Youtube link')}
-              style={{
-                borderRadius: '50%',
-                backgroundColor: '#FF0000',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
-            <Button
-              iconOnly
-              onClick={() => console.log('Bitcoin link')}
-              style={{
-                borderRadius: '50%',
-                backgroundColor: '#F7931A',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
-            <Button
-              iconOnly
-              onClick={() => console.log('Twitter link')}
-              style={{
-                borderRadius: '50%',
-                backgroundColor: '#1DA1F2',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
-            <Button
-              iconOnly
-              onClick={() => console.log('Instagram link')}
-              style={{
-                borderRadius: '50%',
-                background:
-                  'linear-gradient(49.12deg, #FFDD55 6.62%, #FF5D3F 47.35%, #FF543E 50.06%, #C837AB 93.5%)',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
-            <Button
-              iconOnly
-              onClick={() => console.log('Tiktok link')}
-              style={{
-                borderRadius: '50%',
-                backgroundColor: '#000',
-              }}
-            >
-              <ShopCrypto />
-            </Button>
+        <div className="w-full px-[38px] md:p-5 md:mx-auto">
+          <div className="flex flex-col py-3 mb-3 md:border md:border-x-0 md:border-t-0 md:border-b-beerus md:flex-row md:justify-between">
+            <div className="grow-1 mb-6 lg:mr-0 md:mr-4 md:mb-0">
+              <h2 className="text-goten text-sm font-bold mb-7">About Sportsbet.io</h2>
+              <p className="text-xs text-trunks w-full md:w-[148px]">
+                Sportsbet.io is owned and operated by mBet Solutions NV (Schout
+                Bij Nacht Doormanweg 40, P.O. Box 4745, Curaçao). It is licensed
+                and regulated by the Government of Curaçao under the gaming
+                license 1668/JAZ. Some payment methods are handled by its wholly
+                owned subsidiary, mProcessing Solutions Ltd, Cyprus (Menandrou 4,
+                1066, Nicosia, Cyprus).
+              </p>
+            </div>
+            <div className="grow-3 hidden lg:mr-0 md:block md:mr-4">
+              <div>
+                <h2 className="text-goten text-sm font-bold mb-5">Sports</h2>
+                <p className="mb-3 text-trunks text-xs">Promotions</p>
+                <p className="mb-3 text-trunks text-xs">In play</p>
+                <p className="mb-3 text-trunks text-xs">Upcoming</p>
+              </div>
+            </div>
+            <div className="grow-3 hidden lg:mr-0  md:block md:mr-4">
+              <div>
+                <h2 className="text-goten text-sm font-bold mb-5">Casino</h2>
+                <p className="mb-3 text-trunks text-xs">Casino</p>
+                <p className="mb-3 text-trunks text-xs">Live Casino</p>
+                <p className="mb-3 text-trunks text-xs">Bombay club</p>
+              </div>
+            </div>
+            <div className="grow-3 hidden lg:mr-0  md:block md:mr-4">
+              <div>
+                <h2 className="text-goten text-sm font-bold mb-5">Support</h2>
+                <p className="mb-3 text-trunks text-xs">Affiliate FAQ</p>
+                <p className="mb-3 text-trunks text-xs">General Betting Rules</p>
+                <p className="mb-3 text-trunks text-xs">Help Centre</p>
+                <p className="mb-3 text-trunks text-xs">Payment Options</p>
+                <p className="mb-3 text-trunks text-xs">Responsible Gambling</p>
+                <p className="mb-3 text-trunks text-xs">Reward Terms and Conditions</p>
+                <p className="mb-3 text-trunks text-xs">Self-Exclusion Terms and Conditions</p>
+                <p className="mb-3 text-trunks text-xs">The Clubhouse FAQ</p>
+              </div>
+            </div>
+            <div className="grow-3 hidden lg:mr-0  md:block md:mr-4">
+              <div>
+                <h2 className="text-goten text-sm font-bold mb-5">About</h2>
+                <p className="mb-3 text-trunks text-xs">Sportsbet app</p>
+                <p className="mb-3 text-trunks text-xs">Affiliates</p>
+                <p className="mb-3 text-trunks text-xs">Affiliate Terms and Conditions</p>
+                <p className="mb-3 text-trunks text-xs">Crypto Currency How To Guide</p>
+                <p className="mb-3 text-trunks text-xs">Privacy Policy</p>
+                <p className="mb-3 text-trunks text-xs">Sportsbet.io Partnerships</p>
+                <p className="mb-3 text-trunks text-xs">Terms and Conditions</p>
+                <p className="mb-3 text-trunks text-xs">The Clubhouse Terms and Conditions</p>
+                <p className="mb-3 text-trunks text-xs">Vip</p>
+              </div>
+            </div>
+            <div className="grow-4 hidden md:block">
+              <h2 className="text-goten text-sm font-bold mb-5">Preferences</h2>
+              <div className="rounded-[100px] flex flex-col bg-beerus mb-4 ">
+                <div className="flex items-center justify-between my-[6px] ml-3 mr-[17px]">
+                  <div className="flex items-center">
+                    <span className="texst-xs ml-[5px]">English </span>
+                  </div>
+                  <div className="cursor-pointer">
+                    <ControlsChevronDown />
+                  </div>
+                </div>
+              </div>
+              <div className="rounded-[100px] bg-beerus flex justify-center align-middle">
+                <span className="my-[6px] ml-3 mr-[17px]">
+                  <Image src={image1} className="mt-[6px]" />
+                </span>
+              </div>
+            </div>
+          </div>
+          <div className=" hidden py-3 mb-3 justify-between border border-x-0 border-t-0 border-b-beerus md:flex ">
+            <div>
+              <h2 className="text-goten text-sm font-bold mb-5">Payment methods</h2>
+              <div className="flex flex-row  md:w-[70%] sm:flex-wrap">
+                <div className=" bg-[#F7931A] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Bitcasino />
+                </div>
+                <div className=" bg-[#345D9D] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <div className="bg-white rounded-full">
+                    <LtchLogo />
+                  </div>
+                </div>
+                <div className=" bg-[#EC0928] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <TronLogo />
+                </div>
+                <div className=" bg-[#7748FA] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <EthereumIc />
+                </div>
+                <div className=" bg-[#2ABB92] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <UsdIc />
+                </div>
+                <div className=" bg-[#02A6F0] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <XAbove />
+                </div>
+                <div className=" bg-[#0033AD] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Cardano />
+                </div>
+                <div className=" bg-[#0AB758] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <BrazilianIc />
+                </div>
+                <div className=" bg-[#335DF4] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <EuroIcon />
+                </div>
+                <div className=" bg-[#FF3D44] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <JapanesIc />
+                </div>
+                <div className=" bg-[#FFE8BC] rounded-full mr-2 mb-2 flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <IndianIc />
+                </div>
+              </div>
+            </div>
+            <div className="hidden md:block sm:ml-3 sm:mr-3">
+              <h2 className="text-goten text-sm font-bold mb-5">Follow us</h2>
+              <div className="flex flex-row md:flex-nowrap sm:flex-wrap">
+                <div className="bg-[#27A4DB] rounded-full mr-2  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Telegram />
+                </div>
+                <div className="bg-[#FF0000] rounded-full mr-2  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Youtube />
+                </div>
+                <div className="bg-[#F7931A] rounded-full mr-2  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Bitcasino />
+                </div>
+                <div className="bg-[#1DA1F2] rounded-full mr-2  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Twitter />
+                </div>
+                <div className="rounded-full mr-2  flex items-center justify-center bg-gradient-to-r from-yellow-500 to-pink-500 lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Instagram />
+                </div>
+                <div className="bg-[#000000] rounded-full mr-2  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Tiktok />
+                </div>
+              </div>
+            </div>
+
+            <div>
+              <h2 className="text-goten text-sm font-bold mb-5 hidden md:block">
+                Partnership
+              </h2>
+              <div className="flex items-center">
+                <div className="bg-beerus mr-2 rounded-full  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Arsenal />
+                </div>
+                <div className="bg-beerus mr-2 rounded-full  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Southampton />
+                </div>
+                <div className="bg-beerus mr-2 rounded-full  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Flamengo />
+                </div>
+                <div className="bg-beerus mr-2 rounded-full  flex items-center justify-center lg:w-[36px] lg:h-[36px] sm:w-[35px] sm:h-[35px]">
+                  <Watford />
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="flex flex-col md:hidden">
+            <div className=" mb-2">
+              <Accordion
+                title={'List item'}
+                isContentInside={true}
+                backgroundColor='goku'
+              />
+            </div>
+            <div className=" mb-2">
+              <Accordion
+                title={'Casino'}
+                isContentInside={true}
+                backgroundColor='goku'
+              >
+                <div className="flex flex-col">
+                  <p className="mb-3 text-trunks text-xs">Casino</p>
+                  <p className="mb-3 text-trunks text-xs">Live Casino</p>
+                  <p className="mb-3 text-trunks text-xs">Bombay club</p>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-2">
+              <Accordion backgroundColor='goku' title={'Support'} isContentInside={true}>
+                <div className="flex flex-col">
+                  <p className="mb-3 text-trunks text-xs">Affiliate FAQ</p>
+                  <p className="mb-3 text-trunks text-xs">
+                    General Betting Rules
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">Help Centre</p>
+                  <p className="mb-3 text-trunks text-xs">Payment Options</p>
+                  <p className="mb-3 text-trunks text-xs">Responsible Gambling</p>
+                  <p className="mb-3 text-trunks text-xs">
+                    Reward Terms and Conditions
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">
+                    Self-Exclusion Terms and Conditions
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">The Clubhouse FAQ</p>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-2">
+              <Accordion backgroundColor='goku' title={'About'} isContentInside={true}>
+                <div className="flex flex-col">
+                  <p className="mb-3 text-trunks text-xs">Sportsbet app</p>
+                  <p className="mb-3 text-trunks text-xs">Affiliates</p>
+                  <p className="mb-3 text-trunks text-xs">
+                    Affiliate Terms and Conditions
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">
+                    Crypto Currency How To Guide
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">Privacy Policy</p>
+                  <p className="mb-3 text-trunks text-xs">
+                    Sportsbet.io Partnerships
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">Terms and Conditions</p>
+                  <p className="mb-3 text-trunks text-xs">
+                    The Clubhouse Terms and Conditions
+                  </p>
+                  <p className="mb-3 text-trunks text-xs">Vip</p>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-2">
+              <Accordion backgroundColor='goku' title={'Preferences'} isContentInside={true}>
+                <div className="flex flex-col">
+                  <div className="rounded-[100px] flex flex-col bg-beerus mb-4 ">
+                    <div className="flex items-center justify-between my-[6px] ml-3 mr-[17px]">
+                      <div className="flex items-center">
+                        <span className="texst-xs ml-[5px]">English </span>
+                      </div>
+                      <div className="cursor-pointer">
+                        <ControlsChevronDown />
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-2">
+              <Accordion backgroundColor='goku' title={'Accapted Methods'} isContentInside={true}>
+                <div className="flex flex-col ">
+                  <div className="flex w-[70%] flex-wrap items-center ">
+                    <div className=" bg-[#345D9D] rounded-full mr-2 mb-2 flex items-center justify-center w-[40px] h-[40px]">
+                      <XAbove />
+                    </div>
+                    <div className="  bg-[#F7931A] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px] ">
+                      <Bitcasino />
+                    </div>
+                    <div className=" bg-[#345D9D] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px] ">
+                      <div className="bg-white rounded-full">
+                        <LtchLogo />
+                      </div>
+                    </div>
+                    <div className=" bg-[#7748FA] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px] ">
+                      <TronLogo />
+                    </div>
+                    <div className=" bg-[#2ABB92] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px] ">
+                      <BrazilianIc />
+                    </div>
+                    <div className=" bg-black rounded-full mr-2 mb-2  flex items-center justify-center w-[45px] h-[45px] ">
+                      <Mastercard />
+                    </div>
+                    <div className=" bg-[#1A1F71] rounded-full mr-2 mb-2  flex items-center justify-center w-[45px] h-[45px] ">
+                      <Visa />
+                    </div>
+                    <div className=" bg-[#5F1E5E] rounded-full mr-2 mb-2  flex items-center justify-center w-[45px] h-[45px] ">
+                      <Scril />
+                    </div>
+                    <div className=" bg-[#83BA3B] rounded-full mr-2 mb-2  flex items-center justify-center w-[45px] h-[45px] ">
+                      <Neteller />
+                    </div>
+                    <div className=" bg-beerus rounded-full mr-2 mb-2  flex items-center justify-center w-[45px] h-[45px] ">
+                      <PlusFive />
+                    </div>
+                  </div>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-2">
+              <Accordion backgroundColor='goku' title={'Follow Sportsbook.io'} isContentInside={true}>
+                <div className="flex flex-col">
+                  <div className="flex flex-wrap">
+                    <div className="bg-[#27A4DB] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Telegram />
+                    </div>
+                    <div className="bg-[#FF0000] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Youtube />
+                    </div>
+                    <div className="bg-[#F7931A] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Bitcasino />
+                    </div>
+                    <div className="bg-[#1DA1F2] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Twitter />
+                    </div>
+                    <div className="rounded-full mr-2 bg-gradient-to-r from-yellow-500 to-pink-500 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Instagram />
+                    </div>
+                    <div className="bg-[#000000] rounded-full mr-2 mb-2  flex items-center justify-center w-[40px] h-[40px]">
+                      <Tiktok />
+                    </div>
+                  </div>
+                </div>
+              </Accordion>
+            </div>
+            <div className=" mb-5">
+              <Accordion backgroundColor='goku' title={'Sponsorships'} isContentInside={true}>
+                <div className="flex flex-wrap">
+                  <div className="flex flex-wrap">
+                    <div className="bg-beerus rounded-full mr-2 mb-2 flex items-center justify-center w-[40px] h-[40px]">
+                      <Arsenal />
+                    </div>
+                    <div className="bg-beerus rounded-full mr-2 mb-2 flex items-center justify-center w-[40px] h-[40px]">
+                      <Southampton />
+                    </div>
+                    <div className="bg-beerus rounded-full mr-2 mb-2 flex items-center justify-center w-[40px] h-[40px]">
+                      <Flamengo />
+                    </div>
+                    <div className="bg-beerus rounded-full mr-2 mb-2 flex items-center justify-center w-[40px] h-[40px]">
+                      <Watford />
+                    </div>
+                  </div>
+                </div>
+              </Accordion>
+            </div>
+          </div>
+          <div className="flex justify-center mb-6 md:hidden">
+            <div className="rounded-[100px] bg-beerus flex justify-center items-center">
+              <span className="mt-[6px] ml-3 mr-[17px]">
+                <Image src={image1} className="mt-[6px]" />
+              </span>
+            </div>
+          </div>
+          <div>
+            <h2 className="text-goten text-sm font-bold mb-5">Awards</h2>
+            <div className="flex items-center flex-wrap md:flex-nowrap">
+              <div className="mr-2 ">
+                <Image src={Curacao} />
+              </div>
+              <div className="mr-2 ">
+                <Image src={Award1} />
+              </div>
+              <div className="mr-2 ">
+                <Image src={SbcAwards} />
+              </div>
+              <div className="mr-2 ">
+                <Image src={Award2} />
+              </div>
+              <div className="mr-2 ">
+                <Image src={Award3} />
+              </div>
+              <div className="mr-2 ">
+                <Image src={Plus18} />
+              </div>
+            </div>
           </div>
         </div>
       </div>