+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 9.2.2
+
+### Patch Changes
+
+- fix templates
+
 ## 9.2.1
 
 ### Patch Changes