+++ workspaces/templates/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @heathmont/moon-templates
 
+## 9.2.2
+
+### Patch Changes
+
+- fix templates
+
 ## 9.2.1
+
 ### Patch Changes
 
 - fixes